gradient accumulation, mel-generation, bugfixes by @Rick-McCoy
* backup 0826

* backup

* revise tier splitting

* try tier1 first

* update gitignore

* update

* add unblizzard.yaml

* now train runs

* parallelize rnn ops

* fix biGRU

* fix shape

* fix pad

* fix tier split

* use central stack only at initial tier

* add adam optimizer

* view contiguous

* fix tierutil, centralized stack

* gradient accumulation, mel-generation, bugfixes

* Update gitignore.

* Extremely minor update.

* don't use gmm

* train each tiers separately

* validation

* now all tier works

* batch size to args

* a bit hard coded constants

* pad at end, some fixes

* Remove tensorboardX.

* Parallelize mel spec generation.

* use LSTM instead of LSTM

* Removed dependency on tensorboardX.

* normalize spectrograms

* fix norm code

* fix central stack residual connection

* Add upsample.py

* add model/upsample.py

* Added VoxCeleb2 & minor changes.

* Changed reading dataset from scipy to pydub
Now supports almost all formats

* Fixed bug in which .cuda() was done to np.ndarray

* .cuda() cannot be called in dataloader.

* Wrong function name

* Removed wrongly placed .cuda()s in MelGen

* Removed more misplaced .cuda()s in audio.py

* cut_divide_tiers() accounted for nonexistent batch

* Fixed loading in validation.py

* Update .gitignore for dataset

* Add gmm.py

* Fixed minor bug

* Fixed bugs.

* Added SGD optimizer.

* Fixed broadcasting bug

* More bug fixing: .cuda() in get_pi_indices

* Added reshape_as to sample_gmm

* Changed indices to numpy array

* Fix bug in sample_gmm

* Fix typo in validation

* More bug fixing in gmm.py

* Fixed an overlooked bug

* Added mu, std, pi to log_training

* Unconditional w/ GMM, SGD lr=1e-4 (#5)

* backup 0826

* backup

* revise tier splitting

* try tier1 first

* update gitignore

* update

* add unblizzard.yaml

* now train runs

* parallelize rnn ops

* fix biGRU

* fix shape

* fix pad

* fix tier split

* use central stack only at initial tier

* add adam optimizer

* view contiguous

* Update gitignore.

* fix tierutil, centralized stack

* Extremely minor update.

* don't use gmm

* train each tiers separately

* validation

* now all tier works

* batch size to args

* a bit hard coded constants

* pad at end, some fixes

* Remove tensorboardX.

* Parallelize mel spec generation.

* use LSTM instead of LSTM

* Removed dependency on tensorboardX.

* normalize spectrograms

* fix norm code

* fix central stack residual connection

* Add upsample.py

* add model/upsample.py

* Added VoxCeleb2 & minor changes.

* Changed reading dataset from scipy to pydub
Now supports almost all formats

* Fixed bug in which .cuda() was done to np.ndarray

* .cuda() cannot be called in dataloader.

* Wrong function name

* Removed wrongly placed .cuda()s in MelGen

* Removed more misplaced .cuda()s in audio.py

* cut_divide_tiers() accounted for nonexistent batch

* Fixed loading in validation.py

* Update .gitignore for dataset

* Add gmm.py

* Fixed minor bug

* Fixed bugs.

* Added SGD optimizer.

* Fixed broadcasting bug

* More bug fixing: .cuda() in get_pi_indices

* Added reshape_as to sample_gmm

* Changed indices to numpy array

* Fix bug in sample_gmm

* Fix typo in validation

* More bug fixing in gmm.py

* Fixed an overlooked bug

* Added mu, std, pi to log_training

* Integrate upsample.py into tiers 2~N.

* Fixed bug in dataloader
Changed torch.stft to librosa
Added gradient accumulation

* Minor bug fix.

* Changed MelGen to only deal with numpy

* Changed amp_to_db to power_to_db.

* Minor bug fix.

* Fixed major bug
m4a raw data is in 16 bit format, not 8 bit

* dtype error in numpy

* Update README.md

+++ utils/utils.py
@@ -13,7 +13,7 @@ def read_wav_np(wavpath):
     file_format = wavpath.split('.')[-1]
     audio = AudioSegment.from_file(wavpath, file_format)
     data = audio.raw_data
-    wav = np.frombuffer(data, dtype=np.uint8)
+    wav = np.frombuffer(data, dtype=np.int16)
     
     if len(wav.shape) == 2:
         wav = wav[:, 0]
@@ -37,6 +37,6 @@ def cut_wav(L, wav):
                 'constant', constant_values=0.0)
     else:
         start = random.randint(0, samples - L)
-        wav = wav[start:start+L]
+        wav = wav[start:start + L]
 
     return wav
